Unfork student email tests
